move experimental package to module instead
