+++ .ci/test_cabal.sh
@@ -1,10 +1,6 @@
 #!/usr/bin/env bash
 set -xeou pipefail
 
-cabal v2-run clash-protocols:unittests --enable-tests
-cabal v2-run clash-protocols-experimental:unittests --enable-tests
-
-cabal v2-run clash-protocols:doctests --enable-tests
-cabal v2-run clash-protocols-experimental:doctests --enable-tests
-
+cabal v2-run unittests --enable-tests
+cabal v2-run doctests --enable-tests
 cabal v2-sdist all
