--- clash-protocols-experimental/Setup.hs
@@ -1,5 +1,0 @@
-import Distribution.Extra.Doctest (defaultMainWithDoctests)
-import Prelude
-
-main :: IO ()
-main = defaultMainWithDoctests "doctests"
