+++ clash-protocols/tests/Tests/Protocols.hs
@@ -3,22 +3,30 @@ module Tests.Protocols (tests, main) where
 import Test.Tasty
 import Tests.Protocols.Avalon qualified
 import Tests.Protocols.Axi4 qualified
+import Tests.Protocols.BiDf qualified
 import Tests.Protocols.Df qualified
+import Tests.Protocols.Df.Extra qualified
 import Tests.Protocols.DfConv qualified
 import Tests.Protocols.PacketStream qualified
+import Tests.Protocols.ReqResp qualified
 import Tests.Protocols.Vec qualified
 import Tests.Protocols.Wishbone qualified
+import Tests.Protocols.Wishbone.Extra qualified
 
 tests :: TestTree
 tests =
   testGroup
     "Protocols"
-    [ Tests.Protocols.Df.tests
+    [ Tests.Protocols.BiDf.tests
+    , Tests.Protocols.Df.tests
+    , Tests.Protocols.Df.Extra.tests
     , Tests.Protocols.DfConv.tests
     , Tests.Protocols.Avalon.tests
     , Tests.Protocols.Axi4.tests
     , Tests.Protocols.PacketStream.tests
+    , Tests.Protocols.ReqResp.tests
     , Tests.Protocols.Wishbone.tests
+    , Tests.Protocols.Wishbone.Extra.tests
     , Tests.Protocols.Vec.tests
     ]
