+++ clash-protocols/tests/Tests/Protocols/ReqResp.hs
@@ -10,8 +10,9 @@ import Clash.Prelude as C
 import Clash.Hedgehog.Sized.Vector (genVec)
 import Hedgehog (Gen, Property)
 import Protocols
-import Protocols.Hedgehog (defExpectOptions)
-import Protocols.ReqResp as ReqResp
+import Protocols.Experimental.Hedgehog (defExpectOptions)
+import Protocols.Experimental.ReqResp as ReqResp
+import Protocols.Experimental.Simulate
 import Test.Tasty (TestTree)
 import Test.Tasty.Hedgehog.Extra (testProperty)
 import Test.Tasty.TH (testGroupGenerator)
@@ -20,8 +21,8 @@ import Hedgehog qualified as H
 import Hedgehog.Gen qualified as Gen
 import Hedgehog.Range qualified as Range
 import Protocols.Df qualified as Df
-import Protocols.Df.Extra qualified as Df
-import Protocols.Hedgehog qualified as PH
+import Protocols.Experimental.Df.Extra qualified as Df
+import Protocols.Experimental.Hedgehog qualified as PH
 
 smallInt :: Gen Int
 smallInt = Gen.integral (Range.linear 0 10)
