+++ flake.nix
@@ -1,5 +1,5 @@
 {
-  description = "A flake for the clash-protocols, clash-protocols-base and clash-protocols-experimental";
+  description = "A flake for the clash-protocols and clash-protocols-base";
   inputs = {
     clash-compiler.url = "github:clash-lang/clash-compiler";
     circuit-notation = {
@@ -54,10 +54,6 @@
                 root = ./clash-protocols-base;
                 overrides = _: _: final;
               };
-              clash-protocols-experimental = prev.developPackage {
-                root = ./clash-protocols-experimental;
-                overrides = _: _: final;
-              };
             } // circuit-notation.overlays.${system}.${compiler-version} final prev;
           in
             { name = compiler-version; value = overlay; }
@@ -110,7 +106,6 @@
           {
             clash-protocols = hs-pkgs.clash-protocols;
             clash-protocols-base = hs-pkgs.clash-protocols-base;
-            clash-protocols-experimental = hs-pkgs.clash-protocols-experimental;
 
             default = hs-pkgs.${default-package};
           }) all-hs-pkgs;
